Move file system and database to separate module
