+++ database/build.gradle
@@ -8,20 +8,12 @@ repositories {
 }
 
 def versions = [
-        kafka    : project.properties['kafka.version'] ?: "3.3.1",
-        junit    : project.properties['junit.version'] ?: "5.9.1",
-        "commons-net": project.properties['commons-net.version'] ?: "3.8.0",
+        junit: project.properties['junit.version'] ?: "5.9.1",
 ]
 
 dependencies {
     testImplementation "org.junit.jupiter:junit-jupiter:${versions["junit"]}"
     testImplementation project(':it')
-
-    implementation project(':common')
-    implementation "org.apache.kafka:connect-runtime:${versions["kafka"]}"
-    implementation "org.apache.kafka:connect-json:${versions["kafka"]}"
-    implementation "org.apache.kafka:connect-api:${versions["kafka"]}"
-    implementation "commons-net:commons-net:${versions["commons-net"]}"
 }
 
 java {
