+++ database/src/main/java/org/astraea/database/ColumnInfo.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.app.database;
+package org.astraea.database;
 
 import java.util.Objects;
