+++ fs/src/main/java/org/astraea/fs/FileSystem.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.connector.fs;
+package org.astraea.fs;
 
 import java.io.InputStream;
 import java.io.OutputStream;
