+++ fs/src/test/java/org/fs/FtpFileSystemTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.connector.fs;
+package org.astraea.fs;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
